refactor(Graphql): move to module style
