+++ docker-compose-dev.yml
@@ -1,0 +1,20 @@
+version: "3"
+
+services:
+  mongo_data:  
+    image: mongo:3.4
+    container_name: mongo_data
+    command: bin/true
+
+  mongo:  
+    image: mongo:3.4
+    container_name: mongo
+    volumes:
+      - mongo_data:/data/db
+    ports:
+      - "27077:27017"
+    restart: on-failure
+    command: --smallfiles --rest
+
+volumes:
+  mongo_data:
