+++ examples/module/index.js
@@ -1,0 +1,3 @@
+const nap = require('../..');
+
+nap.start();
