--- graphql/custom.js
@@ -1,11 +1,0 @@
-exports.extraUserSchema = {
-  score: Number
-}
-
-exports.extraInstallationSchema = {
-  FCMToken: String
-}
-
-exports.extraAuthenSchema = {
-  isVerified: Boolean
-}
