+++ graphql/models/index.js
@@ -1,0 +1,8 @@
+module.exports = (config) => {
+  return Object.assign(
+   require('./User')(config),
+   require('./Installation')(config),
+   require('./Authen')(config),
+   require('./Error')(config)
+  );
+};
