+++ graphql/resolvers/index.js
@@ -1,0 +1,5 @@
+module.exports = (...args) => {
+  require('./Authen')(...args);
+  require('./Installation')(...args);
+  require('./User')(...args);
+};
